requirements.txt
amqp==2.6.0
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.2.1
Brotli==1.0.9
bs4==0.0.1
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.0
captcha==0.3
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlover==0.3.0
confusable-homoglyphs==3.1.1
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cssselect2==0.4.1
cycler==0.10.0
Cython==0.29.21
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.3
django-aesfield==2.2
django-allauth==0.43.0
django-autoslug==1.9.3
django-bootstrap3==10.0.1
django-configurations==0.8
django-cors-headers==3.4.0
django-credit-cards==0.4.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.9
django-filter==2.3.0
django-flat-responsive==2.0
django-formtools==2.2
django-haystack==3.0
django-heroku==0.3.1
django-image-renderer==0.1.3
django-phone-field==1.8.1
django-picklefield==3.0.1
django-registration==2.4.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sendfile==0.3.11
django-simple-aes-field==0.1.3
django-tables2==2.3.3
django-tagging==0.4.6
djangorestframework==3.11.0
djongo==1.3.3
dlx==1.0.4
docplex==2.15.194
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==4.0.2
fastdtw==0.3.4
fastjsonschema==2.14.4
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Script==2.0.6
future==0.18.2
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.9
imageio==2.6.1
imgaug==0.4.0
inflection==0.5.0
ipaddr==2.2.0
ipykernel==5.4.2
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.6.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-segmentation==0.3.0
kiwisolver==1.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.1
m2secret-py3==1.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-polyfield==5.9
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.4.0
mpmath==1.1.0
msgpack==0.6.2
mysql-connector-python==8.0.22
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.0
networkx==2.5
notebook==6.1.6
ntlm-auth==1.5.0
numpy==1.19.2
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.2.0.32
packaging==20.3
pandas==1.0.3
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.7.0
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.12.0
ply==3.11
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.5
protobuf==3.14.0
psutil==5.7.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pybind11==2.5.0
pyCLI==2.0.3
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.11.1
PyMySQL==0.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyrsistent==0.16.0
pyscf==1.7.4
pysolr==3.9.0
python-constraint==1.4.0
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==20.0.0
qiskit==0.19.6
qiskit-aer==0.5.2
qiskit-aqua==0.7.3
qiskit-ibmq-provider==0.7.2
qiskit-ignis==0.3.3
qiskit-terra==0.14.2
Quandl==3.5.2
reportlab==3.5.42
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retrying==1.3.3
retworkx==0.4.0
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.0
SimpleAES==0.0.0
simplejson==3.17.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
sympy==1.6.1
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tinycss2==1.1.0
toml==0.10.1
tornado==6.1
tqdm==4.42.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
wcwidth==0.2.4
WeasyPrint==52.1
webencodings==0.5.1
websockets==7.0
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xadmin==0.6.1
xhtml2pdf==0.2.4
xlwt==1.3.0
